feat(estatico-boilerplate): extract dependency graph resolvers
